Move the ingress back to the default namespace
Ingress in a KURL install will use the admin console's TLS, so it needs
to be in the same namespace as the admin console to access its secrets.
However it can still provide an ingress for a service in another
namespace.

Update config to clarify how these are separated, and to elaborate on
PFI secret key.