Switch to providing an ingress for the backend service, drop 7000
CD4PE 3.4.0 deprecates the Distelli agent, which was the only thing that
communicated with port 7000. Drop mapping it.

Also drop configuring SSL directly in CD4PE. We'll rely on an encrypted
overlay network and SSL termination on the ingress for it instead.
Eventually cd4pe_jobs tasks need to be updated to follow this model.

Switch to using an ingress to expose the backend service (normally port
8000) to simplify configuration.